requirements.txt
librosa
vosk
sounddevice
dtw-python
PyYAML
RPi.GPIO
